+++ bank-feeds/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [csharp v3.0.0] bank-feeds
 ### Releases
 - [NuGet v3.0.0] https://www.nuget.org/packages/Codat.BankFeeds/3.0.0 - bank-feeds
+
+## 2023-12-14 14:05:02
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.126.3 (2.214.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.1.0] bank-feeds
+### Releases
+- [NuGet v3.1.0] https://www.nuget.org/packages/Codat.BankFeeds/3.1.0 - bank-feeds